requirements.txt
discord.py
python-dotenv
requests
requests-html
bs4